arrayField refactoring
