+++ src/components/common/arrayField/index.ts
@@ -1,2 +1,4 @@
-export type { ArrayFieldProps, ArrayField2Props } from "./arrayField";
-export { default, default as ArrayField, ArrayField2 } from "./arrayField";
+export type { ArrayFieldProps } from "./arrayField";
+export { default, default as ArrayField } from "./arrayField";
+export { ArrayField2 } from "./arrayField2";
+export type { ArrayField2Props } from "./arrayField2";
